added resources and policies

+++ backend-lumen/app/Http/Controllers/Api/UserController.php
@@ -4,55 +4,36 @@ namespace App\Http\Controllers\Api;
 
 use App\Http\Controllers\Controller;
 use Illuminate\Support\Facades\Auth;
-use  App\Entities\User;
+use App\Entities\User;
+use App\Http\Resources\User\User as UserResource;
+use App\Http\Resources\User\UserCollection;
 
 class UserController extends Controller
 {
-    /**
-     * Get the authenticated User.
-     *
-     * @return Response
-     */
-    public function currentUser()
-    {
-        return response()->json([
-            'user' => Auth::user()
-        ], 200);
-    }
 
     /**
      * Get all User.
      *
      * @return Response
-     */
+    */
     public function allUsers()
     {
-         return response()->json([
-             'users' =>  User::all()
-            ], 200);
+        return new UserCollection(User::all());
     }
 
     /**
      * Get one user.
      *
      * @return Response
-     */
+    */
     public function singleUser($id)
     {
-        try {
-            $user = User::findOrFail($id);
+        $user = User::find($id);
 
-            return response()->json([
-                'user' => $user
-            ], 200);
+        if(!$user)
+            return response()->json(['message' => 'user not found!'], 404);
 
-        } catch (\Exception $e) {
-
-            return response()->json([
-                'message' => 'user not found!'
-            ], 404);
-        }
-
+        return new UserResource(User::findOrFail($id));
     }
 
     /**
@@ -60,24 +41,16 @@ class UserController extends Controller
      *
      * @param  User  $user
      * @return \Illuminate\Http\Response
-     */
+    */
     public function deleteUser($id)
     {
-        try {
-            $user = User::findOrFail($id);
+        $user = User::find($id);
 
-            $user->delete();
+        if(!$user)
+            return response()->json(['message' => 'user not found!'], 404);
 
-            return response()->json([
-                'message'  => "User has been successfully deleted.",
-            ], 200);
-
-        } catch (\Exception $e) {
-
-            return response()->json([
-                'message' => 'user not found!'
-            ], 404);
-        }
+        $user->delete();
+        return response()->json(['message'  => "User has been successfully deleted.",], 200);
     }
 
 }
